[refactor] 전체 폴더 위치및 sign 명 join 및 login으로 변경
