+++ src/App.tsx
@@ -1,8 +1,7 @@
 import React from "react";
 import { Outlet } from "react-router-dom";
-import LsbComponent from "./pages/Home/Layout/LsbComponent";
 import { createGlobalStyle } from "styled-components";
-import NavComponent from "./pages/Home/Layout/NavComponent";
+import NavComponent from "./components/Layout/NavComponent";
 
 const GlobalStyle = createGlobalStyle`
   @import url('https://fonts.googleapis.com/css2?family=Source+Sans+Pro:wght@300&display=swap');
