--- src/assets/AssetsDummy.md
@@ -1,1 +1,0 @@
-assets dummy
