--- src/components/ComponentDummy.md
@@ -1,2 +1,0 @@
-
-componentdummy
