+++ src/components/Note/ViewAllNote.tsx
@@ -18,7 +18,7 @@ import { Dispatch, SetStateAction, useEffect, useState } from "react";
 import styled from "styled-components";
 import { Link } from "react-router-dom";
 import PageContainer from "../PageContainer";
-import DashboardCard from "../DashBoardCard";
+import DashboardCard from "../Board/DashBoardCard";
 import { Pagination, Typography } from "@mui/material";
 import { BottomNote } from "./BottomNote";
 import Paging from "../Paging";
