+++ src/pages/Home/Switch/Note.tsx
@@ -1,7 +1,6 @@
 import { AllMiddleNote } from "@/components/Note/AllMiddleNote";
 import { BottomNote } from "@/components/Note/BottomNote";
-import { MiddleNote } from "@/components/Note/MiddleNote";
-import { TopNote } from "@/components/Note/TopNote";
+
 import { Outlet } from "react-router-dom";
 
 function Note() {
@@ -15,20 +14,3 @@ function Note() {
 }
 
 export default Note;
-/*
-
-
-- TopNote
-    - TransferNote
-    - DeleteNote
-- MiddleNote
-    
-       FolderName
-    
-    - SortNote
-        - SortName
-        - SortDate
-- BottomNote
-    - -ViewNote
-    - -MoveBar
-*/
