+++ src/pages/Home/Switch/RefTeamNotes.tsx
@@ -1,6 +1,4 @@
-import { DefaultNotes } from "@/components/Board/DefaultNotes";
 import { MiddleNote } from "@/components/Note/MiddleNote";
-import { TopNote } from "@/components/Note/TopNote";
 import ViewTeamNotes from "@/components/Note/ViewTeamNotes";
 
 function RefTeamNotesPage() {
