+++ src/pages/Home/Switch/ReferenceNote.tsx
@@ -1,7 +1,4 @@
-import { DefaultNotes } from "@/components/Board/DefaultNotes";
 import { AllMiddleNote } from "@/components/Note/AllMiddleNote";
-import { MiddleNote } from "@/components/Note/MiddleNote";
-import { TopNote } from "@/components/Note/TopNote";
 import { ViewAllNotes } from "@/components/Note/ViewAllNote";
 
 function ReferenceNotePage() {
