--- src/pages/PagesDummy.md
@@ -1,1 +1,0 @@
-tests dummy
