--- src/pages/Sign/SignPage.tsx
@@ -1,12 +1,0 @@
-import { SignUp } from "@/components/Sign/SignUp";
-import styled from "styled-components";
-
-function Sign() {
-  return (
-    <div>
-      <SignUp />
-    </div>
-  );
-}
-
-export default Sign;
