--- src/tests/TestsDummy.md
@@ -1,2 +1,0 @@
-
-api dummy
